dashboard: periodically refresh bug subsystems
This will both help set subsystems for bugs that existed before these
changes and keep subsystems up to date with changes to the subsystem
list.

Do the update once a month for open bugs and also update all open and
fixed bugs every time a revision changes.

+++ dashboard/app/config.go
@@ -110,7 +110,7 @@ type SubsystemsConfig struct {
 	// If Service is set, dashboard will use it to infer and recalculate subsystems.
 	Service *subsystem.Service
 	// If all existing subsystem labels must be recalculated, increase this integer.
-	Revision int64
+	Revision int
 }
 
 // ObsoletingConfig describes how bugs without reproducer should be obsoleted.
